flatpak-import: Do metadata comparison in transaction
This ensures that the commit and ref aren't committed before all the
checks are done. One important change is needed. Rather than passing ref
to read_commit, the checksum is passed. Since the ref hasn't been
committed yet, it would either not exist or point to an old commit.
OSTree will look in the staging directory for the commit if it's not in
the objects directory yet and find the to-be-included commit.

https://phabricator.endlessm.com/T17258