flatpak-import: Update repo metadata after import
In order for the repo to be usable before the import task completes,
the metadata needs to be updated with flatpak build-update-repo. This
updates the summary file and appstream branches. No pruning or delta
generation are done here since we want the import task to complete
quickly. It's assumed another process will handle those duties.

https://phabricator.endlessm.com/T17258