ReceiveTask: Always delete upload
It would be nice to keep the upload when the import task fails so it
can be inspected or possibly to save a subsequent upload of the same
file. Unfortunately, nothing cleans up these failed uploads, so they'd
just fill the disk eventually. Likewise, the server throws an
exception at exit trying to remove a non-empty temporary
directory. Just delete them for now.

https://phabricator.endlessm.com/T17258