UploadWebApp: Report back task ID
Send the task ID back in the JSON response for /upload and
/push. Currently nothing interesting can be done with this, but later
it will be used for clients to query a task's state.

There is one glaring flaw, though. The task IDs are generated starting
from 0 each time the server starts, so the IDs are only useful as long
as the server doesn't shut down. This really needs a database to track
the tasks persistently. In the short term, this will be good enough
for a client to make a request, get the task back, and poll it until
completion.

https://phabricator.endlessm.com/T17258