FlatpakImporter: Return early if commits are equal
In addition to checking if the actual commit checksums match, check if
the current ref is built from the uploaded commit. This is done by
looking at the tree content and metadata checksums in the commit
metadata. This will be important soon when the importer is copying the
uploaded commit so that it has the correct collection binding for the
target repository.

https://phabricator.endlessm.com/T19126