FlatpakImporter: Add OSTree experimental API compatibility
In order to support P2P downloads, commits need have valid collection
and ref bindings in their metadata. Unfortunately, the OSTree in use on
SOMA does not have the experimental API, so none of the needed methods
will be available. Add fallbacks for the commit metadata keys as well as
OSTree.Repo.get_collection_id().

https://phabricator.endlessm.com/T19126