FlatpakImporter: Copy bundled commit for correct bindings
The final commit needs to have correct collection and ref bindings on
the server. Since it's almost certainly not the case that the builder
repo is setup like the server, the commit needs to be copied with
corrected binding information in the commit metadata.

The _copy_commit() method is similar to "flatpak build-commit-from"
except we have much more control over how the bindings are generated and
how the transaction is managed. Furthermore, since the flatpak on SOMA
isn't built with --enable-p2p, then it wouldn't fill in those bindings.

https://phabricator.endlessm.com/T19126