Added tests for importing tgz and tarfiles with change to a generic name
Since we will need to be able to import more than just a static delta
bundle as provided by OSTree, we now have a test that can check how the
importing is working with either tgz, tar, or flatpak file. While adding
the tests, filenames/API was changed to reflect this change (without
change in behavior) and linting issues cleaned up.