test: Regenerate test data
One of the keys was expired and the flatpak was signed by a key that no
longer exists. This makes the usage clearer with exported keys for
importing into temporary GPG homedirs.

The flatpak is now built with a current runtime since the 1.6 runtime is
several years old, EOL'd, and unlikely to be installed on the
developer's machine.