Merge flatpak-import.conf into ostree-upload-server.conf
Since the importer is a class loaded by the server rather than a
standalone script like it was originally, the server can provide the
configuration directly to the importer without the need for a separate
configuration file.

This is nicer for deployment since there's only one config file to
manage, but it's also nicer for testing since the all the server
components can be tested without managing a configuration file.