OstreeUploadServer: Gracefully stop on signals
Normally the process is stopped immediately on any terminating signals
and won't reach the cleanup at the end of the run method. Add signal
handlers for SIGTERM and SIGHUP that convert to SystemExit so that the
finally block is reached. The SystemExit isn't caught so that the
message about the signal is printed and and the process exits with a
failure code instead of success. In WorkerPoolExecutor, the exit event
is created during initialization in case the pool is stopped before it's
started.