OstreeUploadServer: Allow running without blocking
Break out the start and stop actions from the run method so that the
server can be started and stopped without blocking in the main loop.
This requires refactoring most of the state into instance attributes,
but it makes actually makes a lot of the operation much clearer.
Unfortunately, the maintenance handling is only done from the blocking
loop until it can be refactored to run from a normal task.