Fix using test configuration in docker
The `test/ostree-upload-server.conf` file is written to expect the test
GPG data and repo in the current directory, but when used in Docker the
file is mounted to the parent directory. This fixes the path to work
correctly in docker but means that manual tests also need to be run from
the parent directory with `-c test/ostree-upload-server.conf`.