Drop Flask-API dependency
I had figured Flask-API was an integral part of the server, but in fact
the only thing used was the `flask_api.status.is_success` utility
function. The entire implementation of that function is:

```
def is_success(code):
    return code >= 200 and code <= 299
```

Uh, I think we can handle that here. This does address an actual issue
where we've forced Flask to the 1.1 series but the latest Flask-API
release claims to only be compatible with Flask 2.